Separate unit tests from integration tests
Split tests/ into tests/unit/ and tests/integration/ directories.
Move conftest.py (S3 fixture) into integration/.
Extract test_all_gather_cat functions into standalone unit test file.
Convert test_clamping to use DummyDatastore instead of MDPDatastore.
Add pytest markers config to pyproject.toml.
