+++ tests/integration/test_plotting.py
@@ -19,8 +19,8 @@ from neural_lam import vis
 from neural_lam.create_graph import create_graph_from_datastore
 from neural_lam.models.graph_lam import GraphLAM
 from neural_lam.weather_dataset import WeatherDataset
-from tests.conftest import init_datastore_example
 from tests.dummy_datastore import DummyDatastore
+from tests.integration.conftest import init_datastore_example
 
 # Create output directory for test figures
 TEST_OUTPUT_DIR = Path(__file__).parent / "test_outputs" / "plotting"
