+++ tests/integration/test_training.py
@@ -13,10 +13,9 @@ from neural_lam import config as nlconfig
 from neural_lam.create_graph import create_graph_from_datastore
 from neural_lam.datastore import DATASTORES
 from neural_lam.datastore.base import BaseRegularGridDatastore
-from neural_lam.models.ar_model import ARModel
 from neural_lam.models.graph_lam import GraphLAM
 from neural_lam.weather_dataset import WeatherDataModule
-from tests.conftest import init_datastore_example
+from tests.integration.conftest import init_datastore_example
 
 
 def run_simple_training(datastore, set_output_std, metrics_watch=None):
@@ -137,67 +136,3 @@ def test_training(datastore_name):
 def test_training_output_std():
     datastore = init_datastore_example("mdp")  # Test only with mdp datastore
     run_simple_training(datastore, set_output_std=True)
-
-
-def test_all_gather_cat_single_device():
-    """
-    Test that all_gather_cat preserves tensor shape on single-device runs.
-    On a single device, all_gather returns the tensor unchanged (no new
-    leading dim), so all_gather_cat should not flatten any existing dims.
-    """
-
-    class MockModule:
-        """Minimal object with mocked single-device all_gather."""
-
-        def all_gather(self, tensor_to_gather, sync_grads=False):
-            # Single-device behavior: return tensor unchanged
-            return tensor_to_gather
-
-    module = MockModule()
-    # Bind the real ARModel.all_gather_cat to our mock
-    module.all_gather_cat = ARModel.all_gather_cat.__get__(module, MockModule)
-
-    # Simulate a 3D metric tensor: (N_eval, pred_steps, d_f)
-    tensor = torch.randn(4, 3, 5)
-    result = module.all_gather_cat(tensor)
-
-    # On single device, shape must be preserved
-    assert result.shape == tensor.shape, (
-        f"all_gather_cat changed shape on single device: "
-        f"{tensor.shape} -> {result.shape}"
-    )
-    assert torch.equal(result, tensor)
-
-
-def test_all_gather_cat_multi_device_simulation():
-    """
-    Test that all_gather_cat correctly flattens when all_gather adds a
-    leading dimension (simulating multi-device behavior).
-    """
-
-    class MockModule:
-        """Object with mocked multi-device all_gather."""
-
-        def all_gather(self, tensor, sync_grads=False):
-            # Simulate 2-GPU all_gather: prepend a dim of size 2
-            return torch.stack([tensor, tensor], dim=0)
-
-    module = MockModule()
-    # Bind the real ARModel.all_gather_cat to our mock
-    module.all_gather_cat = ARModel.all_gather_cat.__get__(module, MockModule)
-
-    tensor = torch.randn(4, 3, 5)  # (N_eval, pred_steps, d_f)
-    result = module.all_gather_cat(tensor)
-
-    # Should flatten (2, 4, 3, 5) -> (8, 3, 5)
-    assert result.shape == (
-        8,
-        3,
-        5,
-    ), f"all_gather_cat wrong shape on multi-device: {result.shape}"
-    # Validate values match expected concatenation along dim 0
-    expected = torch.cat([tensor, tensor], dim=0)
-    assert torch.equal(result, expected), (
-        "all_gather_cat produced incorrectly ordered/combined values "
-        "on multi-device simulation"
-    )
