+++ tests/unit/test_time_slicing.py
